#925 cleanups and refactor

+++ src/TrueOrFalse/Infrastructure/Login/Model/Ops/LoginFromCookie.cs
@@ -2,91 +2,46 @@
 
 public class LoginFromCookie
 {
-    public record struct LoginFromCookieResult(bool Success, string? LoginGuid = null, DateTimeOffset? ExpiryDate = null);
-
-    public static LoginFromCookieResult Run(SessionUser sessionUser,
+    private static void Run(SessionUser sessionUser,
         PersistentLoginRepo persistentLoginRepo,
         UserReadingRepo userReadingRepo,
-        string cookieString)
+        string cookieString,
+        bool? deletePersistentCookie = false)
     {
         var cookieValues = PersistentLoginCookie.GetValues(cookieString);
         if (!cookieValues.Exists())
-        {
-            return new LoginFromCookieResult(false);
-        }
+            throw new Exception("Cookie values do not exist");
 
         var persistentLogin = persistentLoginRepo.Get(cookieValues.UserId, cookieValues.LoginGuid);
         if (persistentLogin == null)
-        {
-            return new LoginFromCookieResult(false);
-        }
+            throw new Exception("Persistent login does not exist");
 
         var user = userReadingRepo.GetById(cookieValues.UserId);
         if (user == null)
-        {
-            return new LoginFromCookieResult(false);
-        }
+            throw new Exception("User does not exist");
 
         sessionUser.Login(user);
 
-        persistentLoginRepo.Delete(persistentLogin);
-        var result = WritePersistentLoginToCookie.Run(cookieValues.UserId, persistentLoginRepo);
-        return new LoginFromCookieResult(true, result.LoginGuid, result.ExpiryDate);
+        if (deletePersistentCookie == true)
+            persistentLoginRepo.Delete(persistentLogin);
     }
 
     public static void Run(SessionUser sessionUser,
         PersistentLoginRepo persistentLoginRepo,
         UserReadingRepo userReadingRepo,
-        string cookieString, HttpContext httpContext
-        )
+        HttpContext httpContext)
     {
-        var cookieValues = PersistentLoginCookie.GetValues(cookieString);
-        if (!cookieValues.Exists())
+        var cookieString = httpContext?.Request.Cookies[PersistentLoginCookie.Key];
+
+        if (cookieString != null && httpContext != null)
         {
-            throw new Exception("Cookie values do not exist");
+            Run(sessionUser, persistentLoginRepo, userReadingRepo, cookieString, deletePersistentCookie: true);
+            WritePersistentLoginToCookie.Run(sessionUser.UserId, persistentLoginRepo, httpContext);
         }
-
-        var persistentLogin = persistentLoginRepo.Get(cookieValues.UserId, cookieValues.LoginGuid);
-        if (persistentLogin == null)
-        {
-            throw new Exception("Persistent login does not exist");
-        }
-
-        var user = userReadingRepo.GetById(cookieValues.UserId);
-        if (user == null)
-        {
-            throw new Exception("User does not exist");
-        }
-
-        sessionUser.Login(user);
-
-        persistentLoginRepo.Delete(persistentLogin);
-        WritePersistentLoginToCookie.Run(cookieValues.UserId, persistentLoginRepo, httpContext);
     }
 
     public static void RunToRestore(SessionUser sessionUser,
         PersistentLoginRepo persistentLoginRepo,
         UserReadingRepo userReadingRepo,
-        string cookieString)
-    {
-        var cookieValues = PersistentLoginCookie.GetValues(cookieString);
-        if (!cookieValues.Exists())
-        {
-            throw new Exception("Cookie values do not exist");
-        }
-
-        var persistentLogin = persistentLoginRepo.Get(cookieValues.UserId, cookieValues.LoginGuid);
-        if (persistentLogin == null)
-        {
-            throw new Exception("Persistent login does not exist");
-        }
-
-        var user = userReadingRepo.GetById(cookieValues.UserId);
-        if (user == null)
-        {
-            throw new Exception("User does not exist");
-        }
-
-        sessionUser.Login(user);
-    }
+        string cookieString) => Run(sessionUser, persistentLoginRepo, userReadingRepo, cookieString, deletePersistentCookie: false);
 }
